Push all the shelf logic down into a standalone file
